+++ projects/scroll-position/src/scroll-position/module/directive/scroll-position.directive.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from "@angular/core";
+import { ScrollPositionDirective } from "./scroll-position.directive";
+
+@NgModule({
+  declarations: [ScrollPositionDirective],
+  exports: [ScrollPositionDirective],
+})
+export class ScrollPositionDirectiveModule {}