+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Unreleased
 
-- none
+- [FEATURE] `IsLoadingService#isLoading$(IGetLoadingOptions)` now accepts an array of keys in addition to a single key argument. If an array of keys if passed, the observable will emit `true` so long as any key is loading, and `false` otherwise.
+- [FEATURE] Added a new `@service-work/scroll-position` package that integrates with the `IsLoadingService` to save and refresh an element's scroll position. The package includes a `ScrollPositionService` that can be used stand-alone as well as a `ScrollPositionDirective` that can be applied to an element to save and automatically refresh an element's scroll position after route navigation.
 
 ### 4.0.0 / 2020-11-11